@@ -206,10 +230,14 @@ subprojects { project ->
     apply plugin: 'maven-publish'
     apply plugin: 'signing'
 
-    sourceCompatibility = "1.11"
-    targetCompatibility = "1.11"
+    if (project.name == "examples-test-data-service") {
+        apply plugin:"org.grails.plugins.views-json"
+    }
 
     java {
+        toolchain {
+            languageVersion = JavaLanguageVersion.of(17)
+        }
         withJavadocJar()
         withSourcesJar()
     }
